スクリプトを分割した
model.ExcelQtConverter、model.RelationProxyModelを独立したモジュールexcelio、relation_proxy_modelにした。これらのモジュールはmodelモジュールに依存しない。今後の開発でmodelに依存するようなコードが紛れ込むと疎結合性が損なわれるので、それを防ぐためにスクリプトを分けて独立性を強調した。またそれに合わせてテストも分割した。
